Move SingleWorker to separate file.
